Fix requested on PR
