+++ backend/src/trainees/trainees.types.ts
@@ -1,0 +1,13 @@
+import { UserArguments } from '../types/UserArguments';
+
+export interface TraineeArguments {
+  user: UserArguments;
+  pesel: string;
+  pkk: string;
+  driversLicenseNumber: string | null;
+}
+
+export interface TraineeArgumentsUpdate
+  extends Partial<Omit<TraineeArguments, 'user'>> {
+  user?: Partial<UserArguments>;
+}
